requirements.txt
botbuilder-integration-aiohttp==4.15.0
aiohttp==3.9.3
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-ai==4.15.0